requirements.txt
datasets>=3.3.2
dvc>=3.59.0
dvc-s3>=3.2.0
peft>=0.14.0
python-dotenv>=1.0.1
trl>=0.15.2
typer>=0.15.1
wandb>=0.19.7

jupyter>=1.1.1
ruff>=0.9.5